Create a 'breed type desync' pattern with Fish reproduction.
Currently whenever a fish gives birth it sets its breed time back to Fish.FishBreedTime and also sets its childs breed time to Fish.FishBreedTime. This makes it so that all breeds happen in sync on the same frame which looks a bit weird. Instead. when a fish is born, its initial BreedTime will be FishBreedTime+1. This will cause a "desync waterfall" so that Fish don't breed in sync every x frames. Instead it is a bit more chaotic.