Fixed Synchronization Issue
There was a big problem with the way the synchronization locking mechanism worked before. Previously, for each point x y in the domain of a thread A, I made sure that there was an Entity e there before entering the mutex lock. This was a big no no. Because what would happen is that while waiting for the mutex lock, a thread B (who A is waiting for to finish processing the critical zone) could process a  Shark to move to e's position in the grid effectively killing it. Then when it was thread A's turn to enter the lock, the Entity e is processed but no longer exists, thus causing an error when attempting to move because EntityMap[e.X, e.Y] is != e. On the contrary it also doesn't make sense to only enter the mutex lock if there's an entity because there's a possibility, that, had we waited for the mutex ock, an Entity would appear there. With all this being said the best course of action would be to enter the mutex lock upon Point2D evaluation, not upon Entity evaluation. This fixes logical errors and runtime errors.